fastapi_ml/app/requirements_fastapi.txt
gunicorn==22.0.0
uvicorn==0.23.2
fastapi==0.110.3

pymongo==4.8.0
motor==3.5.1
slowapi==0.1.9
redis==5.0.7
python-dotenv==1.0.1

accelerate==0.31.0
langchain-core==0.2.9
langchain-community==0.2.5
langchain-huggingface==0.0.3
sentence-transformers==3.0.1
einops==0.8.0
semantic-chunkers==0.0.6
#semantic-router==0.0.46
